Fix note element style for docx
The note template in the docx plugin has a bug. I had to add '/*'
to the end of the match attribute in order to get the rule applied
to all elements inside of the note element. I also had to create a new
Note style in the Normal.docx file.