Prevent floating items overlapping notes
Block items do not wrap around floating objects
by default. This caused a problem with the note
element since it has a background. The float
would overlap the note background, which looks
ugly.

This fix was only applied to the note element since
its the only block element with a background in our
documents. It's possible we may need to apply this
fix to other elements in the future.