Fix issue of ul displaying as ol
When an unordered list was nested in an ordered list, numbers were being placed before each list item instead of bullets. This was because the counter for ordered lists was being applied to all list items nested inside of an ordered list. This fix only applies the counter to list items whose parent is an ordered list.